改为多线程后稳定的版本

+++ LibCore/MsgLib/src/ObjectMsgCall.cpp
@@ -55,7 +55,7 @@ namespace Msg
 
 	CErrno ObjectMsgCall::CopyExcludeParamsAndTargets(ObjectMsgCall *& pMsg ,const std::vector<Msg::Object> & vecTargets , Msg::Object objSrc)
 	{
-		pMsg = new(sizeof(Object)*vecTargets.size())ObjectMsgCall;
+		pMsg = new(sizeof(Object)*(UINT32)(vecTargets.size()))ObjectMsgCall;
 
 		pMsg->m_ullMsgID       = m_ullMsgID;
 		pMsg->m_objSource      = objSrc;
